+++ TimeBot/UserData/UserAccount.cs
@@ -3,7 +3,7 @@
     public class UserAccount
     {
         public ulong userID { get; set; }
-        public int localTime { get; set; }
+        public double localTime { get; set; }
         public string country { get; set; }
     }
 }
